// 1-front/src/app/_services/generated-api/model/signupRequest.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SignupRequest { 
    prenom: string;
    nom: string;
    email: string;
    password: string;
    tel: string;
}