// 1-front/src/app/_services/generated-api/model/userDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserDTO { 
    id?: number;
    nom?: string;
    createdAt?: Date;
    prenom?: string;
    email?: string;
    tel?: string;
    role?: UserDTO.RoleEnum;
}
export namespace UserDTO {
    export type RoleEnum = 'ROLE_USER' | 'ROLE_MANAGER' | 'ROLE_ADMIN' | 'ROLE_SUPERVISEUR';
    export const RoleEnum = {
        RoleUser: 'ROLE_USER' as RoleEnum,
        RoleManager: 'ROLE_MANAGER' as RoleEnum,
        RoleAdmin: 'ROLE_ADMIN' as RoleEnum,
        RoleSuperviseur: 'ROLE_SUPERVISEUR' as RoleEnum
    };
}